simpler prepare script for next gen offline audio

+++ bin/ingest/prepare_v2.sh
@@ -2,12 +2,12 @@
 
 export SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )";
 
-export DATE_CONV_SCRIPT="$SCRIPT_DIR/_iso_to_unix_timestamp.js";
-rm -f $DATE_CONV_SCRIPT;
-echo "#!/usr/bin/env node" > $DATE_CONV_SCRIPT;
-echo "var args = process.argv.slice(2);" >> $DATE_CONV_SCRIPT;
-echo "console.log((new Date(Date.parse(args[0]))).valueOf());" >> $DATE_CONV_SCRIPT;
-chmod a+x $DATE_CONV_SCRIPT;
+# export DATE_CONV_SCRIPT="$SCRIPT_DIR/_iso_to_unix_timestamp.js";
+# rm -f $DATE_CONV_SCRIPT;
+# echo "#!/usr/bin/env node" > $DATE_CONV_SCRIPT;
+# echo "var args = process.argv.slice(2);" >> $DATE_CONV_SCRIPT;
+# echo "console.log((new Date(Date.parse(args[0]))).valueOf());" >> $DATE_CONV_SCRIPT;
+# chmod a+x $DATE_CONV_SCRIPT;
 
 export GUARDIAN_GUID=$1;
 export AUDIO_DIR=$2;
@@ -19,6 +19,10 @@ FILES_PER_DIRECTORY=3990;
 
 FILE_COUNT=0;
 DIR_COUNT=0;
+
+TS_DATE_DELIM="T";
+TS_TIME_DELIM=":";
+DATETIME_ISO_FORMAT="%Y-%m-%dT%H-%M-%S%z"
 
 if [ ! -d $AUDIO_DIR/$GUARDIAN_GUID-$GUARDIAN_NAME ] ; then mkdir -p $AUDIO_DIR/$GUARDIAN_GUID-$GUARDIAN_NAME 2> /dev/null; fi;
 
@@ -38,36 +42,48 @@ for AUDIO_DIR_YEAR_MONTH in $AUDIO_DIR/20*; do
 
         if [ $AUDIO_DIR_TEST != $AUDIO_FILE ] ; then
 
-          TS_DATE_DELIM="T";
-          TS_TIME_DELIM=":";
-          TS_ZONE_DELIM="+";
-          TS_BASE=`echo $AUDIO_FILE | rev | cut -d'_' -f 1 | rev`;
-          TS_DATE=`echo $TS_BASE | cut -d'T' -f 1`;
-          TS_HOUR=`echo $TS_BASE | cut -d'T' -f 2 | cut -d'.' -f 1 | cut -d'-' -f 1`;
-          TS_MIN=`echo $TS_BASE | cut -d'T' -f 2 | cut -d'.' -f 1 | cut -d'-' -f 2`;
-          TS_SEC=`echo $TS_BASE | cut -d'T' -f 2 | cut -d'.' -f 1 | cut -d'-' -f 3`;
-          TS_MILLISEC=`echo $TS_BASE | cut -d'.' -f 2 | cut -d'-' -f 1`;
-          TS_OFFSET=`echo $TS_BASE | cut -d'.' -f 2 | cut -d'-' -f 2`;
-          # TS_MILLISEC_NUMBER=$((TS_MILLISEC+0));
-          # TS_MILLISEC_PADDED=`printf %03d $TS_MILLISEC_NUMBER`;
+          AUDIO_FILE_BASE=$(basename -- "$AUDIO_FILE")
 
-          TS="$TS_DATE$TS_DATE_DELIM$TS_HOUR$TS_TIME_DELIM$TS_MIN$TS_TIME_DELIM$TS_SEC.$TS_MILLISEC$TS_ZONE_DELIM$TS_OFFSET";
+          if (( $((${#AUDIO_FILE_BASE}+0)) > 45 )); then
 
-          if (( $FILE_COUNT % $FILES_PER_DIRECTORY == 0 )) ; then DIR_COUNT=$((DIR_COUNT+1)); fi
+            TS_BASE=`echo $AUDIO_FILE_BASE | rev | cut -d'_' -f 1 | rev`;
+            TS_DATE=`echo $TS_BASE | cut -d'T' -f 1`;
+            TS_HOUR_MIN_SEC=`echo $TS_BASE | cut -d'T' -f 2 | cut -d'.' -f 1`;
 
-          DIR_COUNT_STR=`printf %02d $DIR_COUNT`;
+            TS_MS_AND_ZONE=`echo $TS_BASE | cut -d'.' -f 2 | cut -d'-' -f 1`;
+            TS_MS_TXT="${TS_MS_AND_ZONE:0:3}"
+            if [ "$TS_MS_TXT" = "000" ]; then
+              TS_MS_TXT="0";
+            elif [ "${TS_MS_TXT:0:2}" = "00" ]; then
+              TS_MS_TXT="${TS_MS_TXT:2:1}"
+            elif [ "${TS_MS_TXT:0:1}" = "0" ]; then
+              TS_MS_TXT="${TS_MS_TXT:1:2}"
+            fi
+            TS_MS=$((TS_MS_TXT+0));
 
-          DEST_DIR="$AUDIO_DIR/$GUARDIAN_GUID-$GUARDIAN_NAME/$GUARDIAN_GUID-$DIR_COUNT_STR";
+            read -r DATETIME_ISO DATETIME_EPOCH_ <<< "$(date -jf "$DATETIME_ISO_FORMAT" '+%Y-%m-%dT%H:%M:%S.000%z %s' "$TS_DATE$TS_DATE_DELIM$TS_HOUR_MIN_SEC${TS_MS_AND_ZONE:3:5}")"
 
-          if [ ! -d $DEST_DIR ] ; then mkdir -p $DEST_DIR 2> /dev/null; fi;
+            TS_UNIX=$((DATETIME_EPOCH_*1000+TS_MS))
 
-          export TS_UNIX=`$DATE_CONV_SCRIPT $TS;`;
+            if (( $FILE_COUNT % $FILES_PER_DIRECTORY == 0 )) ; then DIR_COUNT=$((DIR_COUNT+1)); fi
 
-          cp $AUDIO_FILE $DEST_DIR/$TS_UNIX.$AUDIO_FORMAT;
+            DIR_COUNT_STR=`printf %02d $DIR_COUNT`;
 
-          if [ -f $DEST_DIR/$TS_UNIX.$AUDIO_FORMAT ] ; then FILE_COUNT=$((FILE_COUNT+1)); rm $AUDIO_FILE; fi;
+            DEST_DIR="$AUDIO_DIR/$GUARDIAN_GUID-$GUARDIAN_NAME/$GUARDIAN_GUID-$DIR_COUNT_STR";
 
-        fi;
+            if [ ! -d $DEST_DIR ] ; then mkdir -p $DEST_DIR 2> /dev/null; fi;
+
+            cp $AUDIO_FILE $DEST_DIR/$TS_UNIX.$AUDIO_FORMAT;
+
+            if [ -f $DEST_DIR/$TS_UNIX.$AUDIO_FORMAT ] ; then FILE_COUNT=$((FILE_COUNT+1)); rm $AUDIO_FILE; fi;
+
+          else
+
+            echo "file name problem"
+
+          fi
+
+        fi
 
       done
